[조합] 제작시 랜덤 스킬 붙는 로직 고도화

+++ nekoyume/Assets/_Scripts/Game/Skill/AreaAttack.cs
@@ -7,7 +7,7 @@ namespace Nekoyume.Game
     [Serializable]
     public class AreaAttack : AttackSkill
     {
-        public AreaAttack(SkillSheet.Row skillRow, int power, decimal chance) : base(skillRow, power, chance)
+        public AreaAttack(SkillSheet.Row skillRow, int power, int chance) : base(skillRow, power, chance)
         {
         }
